[test] Add retry to `should_be_running` tests
These tests are subject to a timing issue. Since the JvmOptionsParser was introduced
to parse the `jvm.options` before starting logstash, these tests have been subject
to timing issues as the logstash process is the _second_ jvm to run. This commit
stops tests failing if they happen to test for running before the main logstash
process starts